statusLabel moved
